+++ src/main/java/com/lowdragmc/photon/client/gameobject/emitter/particle/ParticleConfig.java
@@ -153,6 +153,7 @@ public class ParticleConfig implements IConfigurable, IPersistedSerializable {
     }
 
     public ParticleConfig() {
+        renderer.getMaterials().add(new MaterialSetting());
     }
 
     @ParametersAreNonnullByDefault