+++ CHANGELOG.md
@@ -7,4 +7,4 @@
 * Added multiple materials rendering supports
 * Added DataFixer for LTS 
 * Added Particle Emitter prewarm
-* 
+* Added Particle Radial Velocity